# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API PGE-QROO",
        "description": "Documentación de la API para la Plataforma de Gestión Energética de Quintana Roo. Permite la gestión de dependencias, edificios, presupuestos y análisis de consumo.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dependencias": {
            "get": {
                "tags": [
                    "Dependencias"
                ],
                "summary": "Listar dependencias",
                "description": "Lista dependencias visibles para el usuario autenticado.",
                "operationId": "5cd8607d5968dd4188851e4fa34ca8a5",
                "responses": {
                    "200": {
                        "description": "Listado de dependencias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_dependencia": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre_dependencia": {
                                                "type": "string",
                                                "example": "Secretaría de Educación"
                                            },
                                            "edificios_count": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dependencias"
                ],
                "summary": "Crear dependencia",
                "operationId": "114ccfc81af4bc0287008030487c9ea6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_dependencia"
                                ],
                                "properties": {
                                    "nombre_dependencia": {
                                        "type": "string",
                                        "example": "Nueva Secretaría"
                                    },
                                    "sector_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dependencia creada"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dependencias/{dependencia}": {
            "get": {
                "tags": [
                    "Dependencias"
                ],
                "summary": "Obtener dependencia especifica",
                "operationId": "a12a8811e351ba31964b4c06acace7ba",
                "parameters": [
                    {
                        "name": "dependencia",
                        "in": "path",
                        "description": "ID de la dependencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la dependencia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_dependencia": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre_dependencia": {
                                            "type": "string",
                                            "example": "Secretaría de Educación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dependencias"
                ],
                "summary": "Actualizar dependencia",
                "operationId": "30198c74ec2f7563e5217eb84081bc13",
                "parameters": [
                    {
                        "name": "dependencia",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_dependencia": {
                                        "type": "string",
                                        "example": "Dependencia Actualizada"
                                    },
                                    "sector_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizada correctamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dependencias"
                ],
                "summary": "Eliminar dependencia",
                "operationId": "bc8c3d70b02beb31113a209e077cc565",
                "parameters": [
                    {
                        "name": "dependencia",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminada correctamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}